feat: compareCard opens cards which will be compared in the evalComparism function
The function compareCard sets up an eventListener to a picked card. The first card will be stored in a variable called firstcard. The secon one in secondCard. Both then will be compared in the functioin evalComparism. If a match has been made, both cards stay open. If not they return to their default position.